적립금 수량의 자료형을 Long으로 변경하라 (#113)

+++ src/main/java/com/kurly/cloud/point/api/order/entity/Order.java
@@ -49,7 +49,7 @@ public class Order {
   int orderProcessCode;
 
   @Column(name = "reserve")
-  int publishPoint;
+  long publishPoint;
 
   @OneToMany(mappedBy = "order")
   @Column(name = "reserve")
