+++ src/main/java/com/kurly/cloud/point/api/point/adapter/in/dto/PublishResultDto.java
@@ -13,7 +13,7 @@ public class PublishResultDto {
   long seq;
   Long memberNumber;
   long orderNumber;
-  Integer charge;
+  Long charge;
   float pointRatio;
   Integer historyType;
   boolean payment;
