+++ src/main/java/com/kurly/cloud/point/api/point/domain/history/PointHistoryInsertRequest.java
@@ -13,7 +13,7 @@ public class PointHistoryInsertRequest {
 
   Long pointSeq;
   long orderNumber;
-  Integer amount;
+  Long amount;
   Integer historyType;
   @Builder.Default
   String detail = "";
