+++ src/main/java/com/kurly/cloud/point/api/point/domain/publish/CancelPublishOrderPointRequest.java
@@ -21,7 +21,7 @@ public class CancelPublishOrderPointRequest {
   @NotNull
   Long orderNumber;
   @NotNull @Min(1)
-  Integer point;
+  Long point;
   @NotNull @Min(0)
   Long actionMemberNumber;
