+++ src/main/java/com/kurly/cloud/point/api/point/exception/NotEnoughPointException.java
@@ -3,7 +3,7 @@ package com.kurly.cloud.point.api.point.exception;
 import java.text.MessageFormat;
 
 public class NotEnoughPointException extends Exception {
-  public NotEnoughPointException(int requested, int has) {
+  public NotEnoughPointException(long requested, long has) {
     super(MessageFormat.format("사용 가능한 적립금이 부족합니다.[가능 : {0}, 요청 : {1}]", has, requested));
   }
 }
